tracker/requirements.txt
psutil
pywin32
pygetwindow
mysql-connector-python
schedule
